firmware,solver,tools,pi: make all Rust clippy-clean
Fix every warning surfaced by the rules_rust clippy aspect
(`--aspects=@rules_rust//rust:defs.bzl%rust_clippy_aspect
--output_groups=clippy_checks`, which runs with `-D warnings`) across all
first-party Rust crates. Lint-only: semantics are preserved exactly; no
heap/alloc, panics, or behavior changes were introduced (the netstack is
no_std firmware on the ESP32-C6).

Idiomatic rewrites: needless_lifetimes, needless_range_loop (only where the
index maps to a single slice), redundant_slicing, manual_memcpy,
unnecessary_cast, needless_return, collapsible_if, write_with_newline,
manual_clamp, manual_range_contains, assign_op_pattern, bool_comparison,
op_ref, len_zero, unnecessary_map_or (is_none_or), int_plus_one,
comparison_chain, approx_constant / excessive_precision (exact f32 consts),
field_reassign_with_default (struct-literal init), needless_update,
redundant_field_names, useless_conversion, non-camel-case type, dead
assignments, and precedence (explicit parens around shifts — grouping
unchanged).

Scoped `#[allow(...)]` with justifying comments only where a rewrite would
hurt clarity/correctness or perturb generated code:
 - matrix/crypto index-arithmetic loops (needless_range_loop in solver, ccmp
   key-schedule/RFC-3394), a deliberate unit error on the AES-unwrap trait
   (result_unit_err), a fixed mathematical fn signature (too_many_arguments),
   conditionally-compiled fields read only on the riscv32 HW-AES backend
   (dead_code), a test's didactic layout arithmetic (identity_op), a test
   input literal near PI (approx_constant);
 - the float validation guards `!(x > 0.0)` in the player
   (neg_cmp_op_on_partial_ord), which intentionally reject NaN — the suggested
   `x <= 0.0` would let a NaN wire value slip past;
 - the fx_vm hot interpreter loop (manual_memcpy, int_plus_one,
   comparison_chain): these run under the fx_bench performance-golden gate, and
   the idiomatic rewrites (copy_from_slice/copy_within, `match x.cmp`) changed
   the emitted RISC-V enough to drift one effect off its golden margin, so the
   original scalar/if-chain forms are kept behind scoped allows. The f32
   math-constant swaps there are bit-identical (verified), so they stay.

missing_safety_doc: converted inline `# Safety:` notes to proper `# Safety`
sections and added truthful sections to the C-ABI FFI entry points.

Note: firmware/netstack/src/tcp.rs is intentionally untouched (handled in a
separate open PR).

All rust_test targets pass; the clippy aspect is warning-free.

+++ firmware/fx_jit/src/lib.rs
@@ -641,7 +641,7 @@ pub mod emu {
                 0x13 => {
                     // I-type ALU (addi sign-extends imm[11:0]; the shifts use shamt)
                     let a = x[rs1];
-                    let shamt = ((w >> 20) & 0x1f) as u32;
+                    let shamt = (w >> 20) & 0x1f;
                     let v = match f3 {
                         0x0 => a.wrapping_add((w as i32) >> 20), // addi
                         0x1 => ((a as u32) << shamt) as i32,     // slli
@@ -745,7 +745,7 @@ mod tests {
     #[test]
     fn jit_matches_reference_over_random_blocks() {
         let mut rng = 0x1234_5678_9abc_def0u64;
-        let consts: Vec<i32> = (0..8).map(|i| (i as i32 - 3) * 37).collect();
+        let consts: Vec<i32> = (0..8).map(|i| (i - 3) * 37).collect();
         for _ in 0..4000 {
             // Build a random VALID straight-line block (tracks depth so ops never
             // underflow), starting from a small random stack.
